+++ frontend/src/common/mock/handlers.ts
@@ -3,18 +3,16 @@ import { editProfileHandlers } from '../../pages/editProfile/mock/handler';
 import { loginHandler } from '../../pages/login/mock/handler';
 import { mentoringCreateHandler } from '../../pages/mentoringCreate/mock/handlers';
 import { mentoringUpdateHandler } from '../../pages/mentoringUpdate/mock/handlers';
-import { signupHandler } from '../../pages/signup/mock/signup/handlers';
-import { validateIdHandler } from '../../pages/signup/mock/validateId/handler';
+import { participatedMentoringHandler } from '../../pages/participatedMentoring/mock/handler';
+import { signupHandler } from '../../pages/signup/mock/handlers';
 
 import { authCodeHandler } from './authCode/authCode';
 import { authCodeVerifyHandler } from './authCodeVerify/authCodeVerify';
 import { commonHandler } from './common/handlers';
 import { mentoringHandler } from './mentoring/handlers';
-import { participatedMentoringHandler } from '../../pages/participatedMentoring/mock/handler';
 
 export const handlers = [
   ...mentoringHandler,
-  ...validateIdHandler,
   ...authCodeHandler,
   ...authCodeVerifyHandler,
   ...signupHandler,
