--- frontend/src/pages/signup/mock/signup/handlers.ts
@@ -1,12 +1,0 @@
-import { http, HttpResponse } from 'msw';
-
-import { API_ENDPOINTS } from '../../../../common/constants/apiEndpoints';
-
-const BASE_URL = process.env.API_BASE_URL;
-const SIGNUP_URL = `${BASE_URL}${API_ENDPOINTS.SIGNUP}`;
-
-const postSignup = http.post(SIGNUP_URL, async () => {
-  return await HttpResponse.json({ message: '회원가입 성공' }, { status: 201 });
-});
-
-export const signupHandler = [postSignup];
